Removed old HilbertTree implementations. Now HilbertTree uses the symmetry class to keep track of orientation. Positions for squares are simply the lower left corners.
Positions are now pulled from tree using a maximum function that uses
position, a width, and a height.

Implemented a maximum function for random circles, a particular
function, and also for drawing a black and white image stored as a 2d
array. The image implmentation could use some fine tuning and maybe a
better algorithm for deciding what level to choose.

Process could be optimized. Right now, entire tree is built, but this is
unnecessary. Also, instead of computing max func, could compute a "do we
continue" function. This could short circuit and save compuation time.